+++ Tests/xcodeprojTests/PBXProj+XCTest.swift
@@ -1,0 +1,9 @@
+import Foundation
+@testable import xcodeproj
+
+extension PBXProj {
+    func encode() -> String {
+        let encoder = PBXProjEncoder()
+        return encoder.encode(proj: self)
+    }
+}