+++ Tests/xcprojTests/PBXProj+XCTest.swift
@@ -1,9 +1,0 @@
-import Foundation
-@testable import xcproj
-
-extension PBXProj {
-    func encode() -> String {
-        let encoder = PBXProjEncoder()
-        return encoder.encode(proj: self)
-    }
-}